Refactor role to work with multiple clusters
The way I am working with this role, is to create multiple pacemaker
clusters across the playbook inventory - the playbook inventory does not
define the cluster as this role assumes.

For my needs I am running this across a fleet of servers, which are
configured into multiple 2-node symmetric pacemaker clusters, so I want
to run this role where the cluster nodes are defined via a variable, and
not taken from the inventory.

Therefore this commit removes a lot of extraneous functionality
(remote_nodes, fence agents) that are surplus to my needs, and refactors
the execution model to rely on new variables 'cluster_nodes' and
'cluster_control_node' variables.

cluster_nodes is a hash containing the hostname+ip details for the
cluster nodes, and cluster_control_node is a boolean that controls on
which nodes the 'pcs_' resources are executed.